api/requirements.txt
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.1.2
django-bootstrap4==1.1.1
django-environ==0.4.5
django-extensions==2.1.3
django-filter==2.0.0
django-macros==0.4.0
django-redis==4.9.0
django-settings-export==1.2.1
djangorestframework==3.8.2
djoser==2.1.0
environ==1.0
idna==2.7
inflection==0.3.1
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.20.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
